+++ test/games/prime/patcher_file_lib/test_item_hints.py
@@ -108,9 +108,7 @@ def test_create_hints_item_detailed(hint_type, empty_patches, pickup, item, loca
         message = "{} can be found in {}.".format(item[1][0].upper() + item[1][1:], location[1])
     elif hint_type == HintType.KEYBEARER:
         message = "The Flying Ing Cache in {} contains {}.".format(location[1], item[1])
-    assert result == [
-        {'asset_id': asset_id, 'strings': [message, '', message]}
-    ]
+    assert result == [{'asset_id': asset_id, 'strings': [message, '', message]}]
 
 @pytest.mark.parametrize("pickup_index_and_guardian", [
     (PickupIndex(43), "&push;&main-color=#FF3333;Amorbis&pop;"),
@@ -148,6 +146,45 @@ def test_create_hints_guardians(empty_patches, pickup_index_and_guardian, pickup
 
     # Assert
     message = f"{guardian} is guarding {item[1]}."
-    assert result == [
-        {'asset_id': asset_id, 'strings': [message, '', message]}
-    ]
+    assert result == [{'asset_id': asset_id, 'strings': [message, '', message]}]
+
+@pytest.mark.parametrize("item", [
+    (HintItemPrecision.DETAILED, "the &push;&main-color=#FF6705B3;Pickup&pop;"),
+    (HintItemPrecision.PRECISE_CATEGORY, "a &push;&main-color=#FF6705B3;movement system&pop;"),
+    (HintItemPrecision.GENERAL_CATEGORY, "a &push;&main-color=#FF6705B3;major upgrade&pop;"),
+    (HintItemPrecision.WRONG_GAME, "the &push;&main-color=#45F731;X-Ray Visor (?)&pop;"),
+])
+@pytest.mark.parametrize("location", [
+    HintLocationPrecision.DETAILED,
+    HintLocationPrecision.WORLD_ONLY,
+    HintLocationPrecision.WRONG_GAME,
+])
+def test_create_hints_light_suit_location(empty_patches, pickup, item, location):
+    # Setup
+    asset_id = 1000
+    pickup_index = PickupIndex(50)
+
+    logbook_node, _, world_list = _create_world_list(asset_id, pickup_index)
+
+    patches = dataclasses.replace(
+        empty_patches,
+        pickup_assignment={
+            pickup_index: pickup,
+        },
+        hints={
+            logbook_node.resource(): Hint(HintType.LIGHT_SUIT_LOCATION,
+                                          PrecisionPair(location, item[0]),
+                                          pickup_index)
+        })
+    rng = MagicMock()
+    rng.choice.side_effect = lambda x: x[0]
+
+    # Run
+    result = item_hints.create_hints(patches, world_list, rng)
+
+    # Assert
+    if location is HintLocationPrecision.WRONG_GAME and item[0] is HintItemPrecision.WRONG_GAME:
+        message = "&push;&main-color=#45F731;Did you remember to check Trial Tunnel?&pop;"
+    else:
+        message = f"U-Mos's reward for returning the Sanctuary energy is {item[1]}."
+    assert result == [{'asset_id': asset_id, 'strings': [message, '', message]}]