+++ CHANGELOG.md
@@ -153,6 +153,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Changed: Each hint for the items guarded by Amorbis, Chykka, and Quadraxis now contains the corresponding
     Guardian's name.
 
+-   Changed: The hint for the vanilla Light Suit location now has special text.
+
 -   Changed: Item names in hints are now colored orange instead of red.
 
 -   Changed: Some joke hints were slightly edited.